Extract config file base test class, use with_env for pumactl tests
